Checkpoint submodule (#21)
