+++ fast_llm/engine/checkpoint/state_dict.py
@@ -1,15 +1,25 @@
+import abc
 import json
 import logging
+import pathlib
+import typing
 
+import safetensors
 import safetensors.torch
 import torch
 import yaml
 
 from fast_llm.core.distributed import safe_barrier
-from fast_llm.engine.config_utils.checkpoint import CheckpointSaveConfig
+from fast_llm.engine.checkpoint.config import CheckpointSaveConfig
 from fast_llm.engine.distributed.distributed import Distributed
+from fast_llm.tensor import SafeTensorSlice
+from fast_llm.utils import Assert
 
 logger = logging.getLogger(__name__)
+
+
+def _import_safetensors_metadata(metadata):
+    return {key: yaml.safe_load(value) for key, value in metadata.items()}
 
 
 def _export_safetensors_metadata(metadata):
@@ -24,6 +34,68 @@ def _export_safetensors_metadata(metadata):
         key: str(value) if isinstance(value, (str, int, float, bool)) else yaml.safe_dump(value)
         for key, value in metadata.items()
     }
+
+
+class StateDictConverter(abc.ABC):
+    base_file_name: typing.ClassVar[str]
+
+    @classmethod
+    @abc.abstractmethod
+    def get_key(cls, parameter_name: str, shard_name: str) -> str:
+        pass
+
+    @abc.abstractmethod
+    def convert_state_dict(
+        self, state_dict: dict[str, torch.Tensor | SafeTensorSlice], export: bool
+    ) -> dict[str, torch.Tensor | SafeTensorSlice]:
+        pass
+
+    @abc.abstractmethod
+    def load_weights(
+        self,
+        directory: pathlib.Path | str,
+        device,
+        shard_names: list[str],
+    ) -> typing.Iterator[tuple[str, str, torch.Tensor | SafeTensorSlice]]:
+        pass
+
+
+class TrivialConverter(StateDictConverter):
+    base_file_name = "state_dict"
+
+    @classmethod
+    def get_key(cls, parameter_name: str, shard_name: str) -> str:
+        return f"{parameter_name}/{shard_name}"
+
+    def convert_state_dict(
+        self, state_dict: dict[str, torch.Tensor | SafeTensorSlice], export: bool
+    ) -> dict[str, torch.Tensor | SafeTensorSlice]:
+        out_state_dict = state_dict.copy()
+        state_dict.clear()
+        return out_state_dict
+
+    def load_weights(
+        self,
+        directory: pathlib.Path | str,
+        device,
+        shard_names: list[str],
+    ) -> typing.Iterator[tuple[str, str, torch.Tensor | SafeTensorSlice]]:
+        index_path = directory / f"state_dict.safetensors.index.json"
+        logger.info(f"Loading index from {index_path}")
+        file_names = set(json.load(index_path.open("r"))["weight_map"].values())
+        for file_name in file_names:
+            logger.info(f"Loading from {directory / file_name}")
+            with safetensors.safe_open(
+                directory / file_name,
+                framework="pt",
+                device=str(device),
+            ) as f:
+                metadata = _import_safetensors_metadata(f.metadata())
+                Assert.eq(metadata["state_shard_names"][: len(shard_names)], list(shard_names))
+                for key in f.keys():
+                    parameter_name, shard_name = key.split("/", 1)
+                    if shard_name in shard_names:
+                        yield parameter_name, shard_name, f.get_slice(key)
 
 
 class StateDictSaver:
