+++ fast_llm/engine/training/config.py
@@ -7,7 +7,7 @@ import typing
 
 from fast_llm.config import Config, Field, FieldHint, FieldUpdate, check_field, config_class, skip_valid_if_none
 from fast_llm.data.config import AbstractDataConfig
-from fast_llm.engine.config_utils.checkpoint import (
+from fast_llm.engine.checkpoint.config import (
     CheckpointConfigBase,
     CheckpointFormat,
     CheckpointLoadConfig,
